+++ tpe/lib/src/Shape.hh
@@ -53,6 +53,12 @@ enum class IGNITION_PHYSICS_TPELIB_VISIBLE ShapeType
 
   /// \brief A mesh shape.
   MESH = 5,
+
+  /// \brief A capsule shape.
+  CAPSULE = 6,
+
+  /// \brief A ellipsoid shape.
+  ELLIPSOID = 7,
 };
 
 
@@ -116,6 +122,49 @@ class IGNITION_PHYSICS_TPELIB_VISIBLE BoxShape : public Shape
   IGN_UTILS_WARN_RESUME__DLL_INTERFACE_MISSING
 };
 
+/// \brief Capsule geometry
+class IGNITION_PHYSICS_TPELIB_VISIBLE CapsuleShape : public Shape
+{
+  /// \brief Constructor
+  public: CapsuleShape();
+
+  /// \brief Copy Constructor
+  /// \param[in] _other shape to copy from
+  public: CapsuleShape(const CapsuleShape &_other);
+
+  /// \brief Destructor
+  public: ~CapsuleShape() = default;
+
+  /// \brief Assignment operator
+  /// \param[in] _other shape to copy from
+  public: Shape &operator=(const Shape &_other);
+
+  /// \brief Get capsule radius
+  /// \return capsule radius
+  public: double GetRadius() const;
+
+  /// \brief Set capsule radius
+  /// \param[in] _radius Cylinder radius
+  public: void SetRadius(double _radius);
+
+  /// \brief Get capsule length
+  /// \return Capsule length
+  public: double GetLength() const;
+
+  /// \brief Set capsule length
+  /// \param[in] _length Cylinder length
+  public: void SetLength(double _length);
+
+  // Documentation inherited
+  protected: virtual void UpdateBoundingBox() override;
+
+  /// \brief Capsule radius
+  private: double radius = 0.0;
+
+  /// \brief Capsule length
+  private: double length = 0.0;
+};
+
 /// \brief Cylinder geometry
 class IGNITION_PHYSICS_TPELIB_VISIBLE CylinderShape : public Shape
 {
@@ -153,6 +202,40 @@ class IGNITION_PHYSICS_TPELIB_VISIBLE CylinderShape : public Shape
 
   /// \brief Cylinder length
   private: double length = 0.0;
+};
+
+/// \brief Ellipsoid geometry
+class IGNITION_PHYSICS_TPELIB_VISIBLE EllipsoidShape : public Shape
+{
+  /// \brief Constructor
+  public: EllipsoidShape();
+
+  /// \brief Copy Constructor
+  /// \param[in] _other shape to copy from
+  public: EllipsoidShape(const EllipsoidShape &_other);
+
+  /// \brief Destructor
+  public: ~EllipsoidShape() = default;
+
+  /// \brief Assignment operator
+  /// \param[in] _other shape to copy from
+  public: Shape &operator=(const Shape &_other);
+
+  /// \brief Get ellipsoid radius
+  /// \return ellipsoid radius
+  public: math::Vector3d GetRadii() const;
+
+  /// \brief Set ellipsoid radius
+  /// \param[in] _radius ellipsoid radius
+  public: void SetRadii(const math::Vector3d &_radii);
+
+  // Documentation inherited
+  protected: virtual void UpdateBoundingBox() override;
+
+  IGN_UTILS_WARN_IGNORE__DLL_INTERFACE_MISSING
+  /// \brief ellipsoid radius
+  private: math::Vector3d radii = math::Vector3d::Zero;
+  IGN_UTILS_WARN_RESUME__DLL_INTERFACE_MISSING
 };
 
 /// \brief Sphere geometry